pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "EnginBert"
version = "0.0.2"
authors = [
    { name = "Jan Heimann", email = "jan_heimann@icloud.com" },
]
description = "A domain-specific BERT model for engineering literature"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
dependencies = [
    "typer[all]",
    "numpy",
    "matplotlib",
    "natsort",
    "pandas",
    "tqdm",
    "fuzzywuzzy",
    "transformers>=4.30.0",
    "torch>=2.0.0",  # Adjust based on your system's compatibility
    "scikit-learn>=1.0",
    "pdfminer.six>=20221105",
]

[project.optional-dependencies]
dev = [
    "requests<2.32.0",
    "pytest>=8.1.1",
    "syrupy>=4.6.1",
    "docker>=7.0.0",
    "flake8>=7.0.0",
    "flake8-pyproject>=1.2.3",
    "black>=24.3.0",
]
docs = [
    "Sphinx>=7.2.6",
    "furo>=2023.9.10",
    "sphinxcontrib-plantuml>=0.27",
    "sphinx-autobuild==2024.4.16",
    "plantuml>=0.3.0",
    "sphinx-copybutton>=0.5.2",
]

[tool.setuptools]
packages = ["scripts"]  # Only include existing directories here

[tool.black]
line-length = 80

[tool.flake8]
max-line-length = 80
extend-ignore = ["E203"]

[project.scripts]
rocketOptimizer = "rocket_optimizer.__main__:main"

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]

